Cache->get option to ignore cache expiration Used by flag on DataService

+++ Cumula/Components/Cache/Cache.php
@@ -58,7 +58,12 @@ class Cache extends \Cumula\Application\SimpleComponent
 	 **/
 	public function get($cacheName, array $options = array()) 
 	{
-		$options = array_merge(array('bin' => 'cache'), $options);
+		$options = array_merge(
+			array(
+				'bin' => 'cache',
+				'ignoreExpiration' => false,
+				),
+			$options);
 		$this->dispatch('cache_populate_datastores');
 		$ds = $this->dataProviders[$options['bin']];
 		$cache = $ds->get($cacheName);
@@ -66,10 +71,11 @@ class Cache extends \Cumula\Application\SimpleComponent
 		if($cache) {
 			$cache = (object)$cache;
 		}
-		if($cache && isset($cache->data) && $cache->expire > time())
-			$return = unserialize($cache->data);
-		if($cache && $cache->expire < time())
+		if($cache && $cache->expire < time() && !$options['ignoreExpiration']) {
 			$ds->destroy($cache);
+		} else if($cache && isset($cache->data)) {
+			$return = unserialize($cache->data);
+		}
 		return $return;
 	} // end function get
